Rename NorentScaffolding to OnboardingScaffolding in backend.
