+++ evictionfree/tests/test_schema.py
@@ -16,7 +16,7 @@ from onboarding.tests.test_schema import (
     _exec_onboarding_step_n,
     exec_legacy_onboarding_step_1,
 )
-from norent.schema import update_scaffolding, SCAFFOLDING_SESSION_KEY
+from onboarding.scaffolding import update_scaffolding, SCAFFOLDING_SESSION_KEY
 from users.tests.factories import UserFactory
 from onboarding.tests.factories import OnboardingInfoFactory
 from evictionfree.tests.factories import (
